requirements.txt
requests>=2
beautifulsoup4>=4
pillow>=10
screeninfo
numpy